refactor(tui): wrap dump in tui.plot following make_plot style
Expose a `dump` wrapper in tui/plot.py paralleling `make_plot`, so the
browser imports it relatively from `.plot` rather than reaching into
`uproot_browser.plot` directly. The width formula is shared with
make_plot and the expr is appended in one place.
